refactor: 전체 배너와 커스텀 배너 사이의 관심사 분리
