+++ src/main/java/com/techeer/fmstudio/domain/banner/service/CustomBannerService.java
@@ -4,19 +4,14 @@ import com.techeer.fmstudio.domain.banner.dao.BannerRepository;
 import com.techeer.fmstudio.domain.banner.dao.MyBannerListRepository;
 import com.techeer.fmstudio.domain.banner.domain.BannerEntity;
 import com.techeer.fmstudio.domain.banner.domain.BannerType;
-import com.techeer.fmstudio.domain.banner.domain.MyBannerList;
 import com.techeer.fmstudio.domain.banner.dto.mapper.BannerMapper;
-import com.techeer.fmstudio.domain.banner.dto.request.CustomBannerAddMyBannerRequest;
 import com.techeer.fmstudio.domain.banner.dto.request.CustomBannerCreateRequest;
 import com.techeer.fmstudio.domain.banner.dto.request.CustomBannerUpdateRequest;
-import com.techeer.fmstudio.domain.banner.dto.response.BannerPageInfo;
 import com.techeer.fmstudio.domain.banner.exception.NotFoundBannerException;
 import com.techeer.fmstudio.domain.member.dao.MemberRepository;
 import com.techeer.fmstudio.domain.member.domain.MemberEntity;
 import com.techeer.fmstudio.domain.member.exception.NotFoundMemberException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -58,13 +53,6 @@ public class CustomBannerService {
                 .orElseThrow(NotFoundBannerException::new);
     }
 
-    @Transactional(readOnly = true)
-    public BannerPageInfo getWholeBannerByPagination(int page, int size){
-        PageRequest pageRequest = PageRequest.of(page, size);
-        Page<BannerEntity> pageResult = bannerRepository.findWholeBannerWithPagination(pageRequest);
-        return bannerMapper.mapEntityToBannerPageInfo(pageResult, page, size);
-    }
-
     public String update(CustomBannerUpdateRequest request, Long id) {
         BannerEntity foundBanner = bannerRepository.findById(id)
                 .orElseThrow(NotFoundBannerException::new);
@@ -87,15 +75,5 @@ public class CustomBannerService {
         foundBanner.deleteBanner();
 
         return foundBanner;
-    }
-
-    public MyBannerList addMyBanner(CustomBannerAddMyBannerRequest request, Long bannerId){
-        MemberEntity member = memberRepository.findMemberEntityByNickname(request.getNickname()).orElseThrow(NotFoundMemberException::new);
-        BannerEntity banner = bannerRepository.findById(bannerId).orElseThrow(NotFoundBannerException::new);
-        MyBannerList myBannerList = MyBannerList.builder()
-                .banner(banner)
-                .member(member)
-                .build();
-        return myBannerListRepository.save(myBannerList);
     }
 }
